Move parts to part to aviod gitignore
