+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/configs/settings.py
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'corsheaders',
-    'vision_on_edge.parts',
+    'vision_on_edge.part',
     'vision_on_edge.streams',
     'vision_on_edge.azure_settings',
     'vision_on_edge.locations',
