+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/azure_training/api/views.py
@@ -27,7 +27,7 @@ from configs.settings import DEVICE_ID, IOT_HUB_CONNECTION_STRING, MODULE_ID
 
 from ...cameras.models import Camera
 from ...general import error_messages
-from ...parts.models import Part
+from ...part.models import Part
 from ..models import Image, Project, Task, Train
 from .serializers import (ImageSerializer, ProjectSerializer, TaskSerializer,
                           TrainSerializer)
