+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/part/__init__.py
@@ -1,0 +1,4 @@
+"""
+vision_on_edge.part init
+"""
+default_app_config = 'vision_on_edge.part.apps.PartsConfig'
