+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/part/admin.py
