+++ factory-ai-vision/EdgeSolution/modules/WebModule/backend/vision_on_edge/part/api/views.py
